feat(navigation): refactor bottom navigation from 4 to 3 items
- Replace HISTORY and MORE destinations with SEARCH and INBOX
- Rename catalog module to search for better semantic clarity
- Create new inbox feature module with beautiful message interface
- Add animated message cards with filtering and search capabilities
- Update all navigation references and routing logic
- Remove deprecated MockNavItems and update navigation structure
- Add proper string resources for new destinations

Key Changes:
- TopLevelDestination: HOME, SEARCH, INBOX (removed HISTORY, MORE)
- New inbox module: Complete MVI implementation with mock data
- Search module: Renamed from catalog with updated navigation
- Clean navigation architecture following modular patterns
