--- app/src/main/java/com/qodein/qode/navigation/MockNavItems.kt
@@ -1,101 +1,0 @@
-package com.qodein.qode.navigation
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import kotlinx.serialization.Serializable
-
-@Serializable
-object HistoryBaseRoute
-
-@Serializable
-object HistoryRoute
-
-@Serializable
-object MoreBaseRoute
-
-@Serializable
-object MoreRoute
-
-@Composable
-fun HistoryScreen(modifier: Modifier = Modifier) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Text(
-            text = "📋 History Screen",
-            style = MaterialTheme.typography.headlineMedium,
-            textAlign = TextAlign.Center,
-        )
-        Text(
-            text = "View your promo code history",
-            style = MaterialTheme.typography.bodyLarge,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(top = 8.dp),
-        )
-    }
-}
-
-fun NavController.navigateToHistory() {
-    navigate(HistoryRoute)
-}
-
-fun NavGraphBuilder.historySection() {
-    composable<HistoryBaseRoute> {
-        HistoryScreen()
-    }
-    composable<HistoryRoute> {
-        HistoryScreen()
-    }
-}
-
-@Composable
-fun MoreScreen(modifier: Modifier = Modifier) {
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Text(
-            text = "⚙️ More Screen",
-            style = MaterialTheme.typography.headlineMedium,
-            textAlign = TextAlign.Center,
-        )
-        Text(
-            text = "Settings and additional options",
-            style = MaterialTheme.typography.bodyLarge,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(top = 8.dp),
-        )
-    }
-}
-
-fun NavController.navigateToMore() {
-    navigate(MoreRoute)
-}
-
-fun NavGraphBuilder.moreSection() {
-    composable<MoreBaseRoute> {
-        MoreScreen()
-    }
-    composable<MoreRoute> {
-        MoreScreen()
-    }
-}
