+++ app/src/main/java/com/qodein/qode/navigation/TopLevelDestination.kt
@@ -2,19 +2,18 @@ package com.qodein.qode.navigation
 
 import androidx.annotation.StringRes
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Menu
+import androidx.compose.material.icons.filled.Inbox
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.outlined.FavoriteBorder
 import androidx.compose.material.icons.outlined.Home
-import androidx.compose.material.icons.outlined.Menu
+import androidx.compose.material.icons.outlined.Inbox
 import androidx.compose.material.icons.outlined.Search
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.qodein.feature.auth.navigation.AuthBaseRoute
-import com.qodein.feature.catalog.navigation.CatalogBaseRoute
-import com.qodein.feature.catalog.navigation.CatalogRoute
 import com.qodein.feature.home.navigation.HomeBaseRoute
+import com.qodein.feature.inbox.navigation.InboxBaseRoute
+import com.qodein.feature.inbox.navigation.InboxRoute
+import com.qodein.feature.search.navigation.SearchBaseRoute
+import com.qodein.feature.search.navigation.SearchRoute
 import com.qodein.qode.R
 import kotlin.reflect.KClass
 import com.qodein.feature.home.R as homeR
@@ -35,28 +34,20 @@ enum class TopLevelDestination(
         route = HomeBaseRoute::class,
         baseRoute = HomeBaseRoute::class,
     ),
-    CATALOG(
+    SEARCH(
         selectedIcon = Icons.Filled.Search,
         unSelectedIcon = Icons.Outlined.Search,
-        iconTextId = R.string.catalog_title,
-        titleTextId = R.string.catalog_title,
-        route = CatalogRoute::class,
-        baseRoute = CatalogBaseRoute::class,
+        iconTextId = R.string.search_title,
+        titleTextId = R.string.search_title,
+        route = SearchRoute::class,
+        baseRoute = SearchBaseRoute::class,
     ),
-    HISTORY(
-        selectedIcon = Icons.Filled.Favorite,
-        unSelectedIcon = Icons.Outlined.FavoriteBorder,
-        iconTextId = R.string.history_title,
-        titleTextId = R.string.history_title,
-        route = HistoryRoute::class,
-        baseRoute = HistoryBaseRoute::class,
-    ),
-    MORE(
-        selectedIcon = Icons.Filled.Menu,
-        unSelectedIcon = Icons.Outlined.Menu,
-        iconTextId = R.string.more_title,
-        titleTextId = R.string.more_title,
-        route = AuthBaseRoute::class,
-        baseRoute = AuthBaseRoute::class,
+    INBOX(
+        selectedIcon = Icons.Filled.Inbox,
+        unSelectedIcon = Icons.Outlined.Inbox,
+        iconTextId = R.string.inbox_title,
+        titleTextId = R.string.inbox_title,
+        route = InboxRoute::class,
+        baseRoute = InboxBaseRoute::class,
     )
 }
