--- feature/catalog/build.gradle.kts
@@ -1,67 +1,0 @@
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.kotlin.compose)
-    alias(libs.plugins.kotlin.serialization)
-    alias(libs.plugins.hilt)
-    alias(libs.plugins.ksp)
-}
-
-android {
-    namespace = "com.qodein.feature.catalog"
-    compileSdk = 36
-
-    defaultConfig {
-        minSdk = 29
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro",
-            )
-        }
-    }
-    kotlin {
-        jvmToolchain(11)
-    }
-
-    buildFeatures {
-        compose = true
-    }
-}
-
-dependencies {
-// Project modules
-    implementation(projects.core.designsystem)
-    implementation(projects.core.ui)
-    implementation(projects.core.model)
-    implementation(projects.core.domain)
-    implementation(projects.core.data)
-
-    // Core Android & Compose
-    implementation(libs.bundles.androidx.core)
-    implementation(platform(libs.androidx.compose.bom))
-    implementation(libs.bundles.compose)
-    implementation(libs.bundles.lifecycle)
-    implementation(libs.bundles.navigation)
-
-    // Dependency Injection
-    implementation(libs.bundles.hilt)
-    ksp(libs.hilt.compiler)
-
-    // Country Code Picker
-    implementation(libs.xmaterial.ccp)
-
-    // Testing
-    testImplementation(libs.bundles.testing.unit)
-    androidTestImplementation(libs.bundles.testing.android)
-
-    // Debug Tools
-    debugImplementation(libs.bundles.debug)
-}
