+++ feature/inbox/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
     alias(libs.plugins.kotlin.serialization)
@@ -8,17 +8,14 @@ plugins {
 }
 
 android {
-    namespace = "com.qodein.qode"
+    namespace = "com.qodein.feature.inbox"
     compileSdk = 36
 
     defaultConfig {
-        applicationId = "com.qodein.qode"
         minSdk = 29
-        targetSdk = 36
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
@@ -30,9 +27,14 @@ android {
             )
         }
     }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
+    }
     kotlin {
         jvmToolchain(11)
     }
+
     buildFeatures {
         compose = true
     }
@@ -41,13 +43,11 @@ android {
 dependencies {
     // Project modules
     implementation(projects.core.designsystem)
-    implementation(projects.feature.auth)
-    implementation(projects.feature.home)
-    implementation(projects.feature.search)
-    implementation(projects.feature.inbox)
+    implementation(projects.core.ui)
     implementation(projects.core.model)
+    implementation(projects.core.domain)
 
-    // Core Android & Compose (using bundles)
+    // Core Android & Compose
     implementation(libs.bundles.androidx.core)
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.bundles.compose)
@@ -61,7 +61,6 @@ dependencies {
     // Testing
     testImplementation(libs.bundles.testing.unit)
     androidTestImplementation(libs.bundles.testing.android)
-    androidTestImplementation(platform(libs.androidx.compose.bom))
 
     // Debug Tools
     debugImplementation(libs.bundles.debug)
