+++ feature/inbox/src/main/AndroidManifest.xml
