+++ feature/inbox/src/main/java/com/qodein/feature/inbox/navigation/InboxNavigation.kt
@@ -1,0 +1,26 @@
+package com.qodein.feature.inbox.navigation
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import com.qodein.feature.inbox.InboxScreen
+import kotlinx.serialization.Serializable
+
+@Serializable
+object InboxBaseRoute
+
+@Serializable
+object InboxRoute
+
+fun NavController.navigateToInbox() {
+    navigate(InboxRoute)
+}
+
+fun NavGraphBuilder.inboxSection() {
+    composable<InboxBaseRoute> {
+        InboxScreen()
+    }
+    composable<InboxRoute> {
+        InboxScreen()
+    }
+}
