+++ settings.gradle.kts
@@ -38,4 +38,5 @@ include(":feature:home")
 include(":core:model")
 include(":core:data")
 include(":core:domain")
-include(":feature:catalog")
+include(":feature:search")
+include(":feature:inbox")
